requirements.txt
aiofiles==0.6.0
aiohttp==3.7.3
APScheduler==3.6.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2020.11.8
cffi==1.14.5
cfscrape==2.1.1
chardet==3.0.4
cloudscraper==1.2.58
cryptography==3.4.5
DateTime==4.3
h11==0.9.0
httpcore==0.11.1
httptools==0.1.1
httpx==0.15.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
lxml==4.6.1
mccabe==0.6.1
multidict==5.0.0
pycparser==2.20
pylint==2.6.0
pyparsing==2.4.7
pyTelegramBotAPI==3.7.6
python-dotenv==0.15.0
python-telegram-bot==13.7
pytz==2020.4
requests==2.25.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
sanic==20.9.1
selenium==3.141.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.0.1
toml==0.10.2
tornado==6.1
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==4.0.1
urllib3==1.26.2
uvloop==0.14.0
websockets==8.1
wrapt==1.12.1
yarl==1.6.3
zope.interface==5.2.0
